Add a simple cli bin crate
For now we query chain state
