+++ src/cli_commands/mod.rs
@@ -14,13 +14,22 @@
 // You should have received a copy of the GNU General Public License along with
 // P2Poolv2. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod bitcoind_rpc;
-pub mod command;
-pub mod config;
-pub mod node;
-pub mod shares;
-pub mod utils;
+pub mod info;
 
-// For testing - make this module public and available regardless of test cfg
-#[cfg(any(test, feature = "testing"))]
-pub mod test_utils;
+// Re-export the shared store functionality
+pub mod store {
+    use p2poolv2_lib::shares::store::Store;
+    use std::error::Error;
+    use std::path::PathBuf;
+
+    /// Open a store from the given path
+    pub fn open_store(store_path: &PathBuf) -> Result<Store, Box<dyn Error>> {
+        println!("Opening store in read-only mode: {:?}", store_path);
+
+        let path_str = store_path.to_str().expect("Invalid path").to_string();
+        Store::new(path_str).map_err(|e| {
+            println!("Failed to open store: {}", e);
+            e
+        })
+    }
+}
